Revert "Minor tweaks to try and fix memory issues" This reverts commit 241ec967b3409a23954ec2ece36000ec9f45fef7. In attempt to fix Segfault issue when building on Linux (Ubuntu).

+++ src/defs.h
@@ -161,7 +161,7 @@ typedef struct {
 	unsigned long long starttime;
 	unsigned long long stoptime;
 	uint8_t depth;
-	uint8_t timeset;
+	long timeset;
 	uint16_t movestogo;
 
 	unsigned long nodes;
